refactor(rust-bundler-manifest-v1): modularize manifest builder internals
