+++ .pr-body-rust-bundler-manifest-modularization-v1.md
@@ -1,0 +1,20 @@
+## Summary
+- refactored `crates/rust-bundler-manifest-v1` into focused internal modules:
+  - `contract.rs` for data contracts and error type
+  - `parsing.rs` for manifest parsing
+  - `validation.rs` for bundle/type/link validation helpers
+  - `ordering.rs` for deterministic ordering/hash helpers
+  - `diagnostics.rs` for diagnostic construction and sorting
+  - `builder.rs` for manifest assembly orchestration
+- reduced `src/lib.rs` to crate glue only (`mod` wiring + `pub use` exports)
+- preserved public API and behavior exactly (same exported types/functions and same test expectations)
+
+## Validation
+- `pnpm install --frozen-lockfile`
+- `pnpm run lint`
+- `pnpm run format:check`
+- `pnpm run build`
+- `pnpm run test`
+- `cargo fmt --all --check`
+- `cargo clippy --workspace --all-targets -- -D warnings`
+- `cargo test --workspace --all-targets`
