+++ crates/rust-bundler-manifest-v1/src/ordering.rs
@@ -1,0 +1,31 @@
+use std::collections::BTreeSet;
+
+use crate::DiagnosticLevel;
+
+pub fn sorted_unique(values: Vec<String>) -> Vec<String> {
+    let mut set = BTreeSet::new();
+    for value in values {
+        set.insert(value);
+    }
+    set.into_iter().collect()
+}
+
+pub fn level_rank(level: &DiagnosticLevel) -> u8 {
+    match level {
+        DiagnosticLevel::Error => 0,
+        DiagnosticLevel::Warn => 1,
+        DiagnosticLevel::Info => 2,
+    }
+}
+
+pub fn stable_hash_64(bytes: &[u8]) -> u64 {
+    const FNV_OFFSET: u64 = 0xcbf29ce484222325;
+    const FNV_PRIME: u64 = 0x100000001b3;
+
+    let mut hash = FNV_OFFSET;
+    for byte in bytes {
+        hash ^= *byte as u64;
+        hash = hash.wrapping_mul(FNV_PRIME);
+    }
+    hash
+}
