+++ crates/rust-bundler-manifest-v1/src/parsing.rs
@@ -1,0 +1,5 @@
+use crate::{ArtifactManifest, ManifestError};
+
+pub fn parse_manifest(raw: &str) -> Result<ArtifactManifest, ManifestError> {
+    Ok(serde_json::from_str(raw)?)
+}
